requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# Mercado Livre Integration
requests==2.31.0
python-decouple==3.8

# Image handling
Pillow==10.2.0

# Utils
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1

# Development
django-debug-toolbar==4.2.0
